refactor(safety): share command policy parser
